KS-2295: Private message: show code error when delete message in special case

+++ eXoApplication/forum/webapp/src/main/java/org/exoplatform/forum/webui/popup/UIForumUserSettingForm.java
@@ -348,7 +348,7 @@ public class UIForumUserSettingForm extends BaseForumForm implements UIPopupComp
 	
 	@SuppressWarnings("unchecked")
   public List<Watch> getListWatch() throws Exception {
-		long pageSelect = pageIterator.getPageSelected() ;
+		int pageSelect = pageIterator.getPageSelected() ;
 		List<Watch>list = new ArrayList<Watch>();
 		try {
 			list.addAll(this.pageList.getPageWatch(pageSelect, this.listWatches)) ;
