+++ eXoApplication/forum/webapp/src/main/java/org/exoplatform/forum/webui/popup/UIViewTopic.java
@@ -70,7 +70,7 @@ public class UIViewTopic extends UIForm implements UIPopupComponent {
 	private Topic topic ;
 	private JCRPageList pageList ;
 	private UserProfile userProfile ;
-	private long pageSelect ;
+	private int pageSelect ;
 	private Map<String, UserProfile> mapUserProfile = new HashMap<String, UserProfile>();
 	RenderHelper renderHelper = new RenderHelper();
 	public UIViewTopic() throws Exception {
@@ -108,7 +108,7 @@ public class UIViewTopic extends UIForm implements UIPopupComponent {
 		pageList = forumService.getPosts(id[l-3], id[l-2], topic.getId(), "", "", "", userLogin)	; 
 		long maxPost = this.userProfile.getMaxPostInPage() ;
 		if(maxPost <= 0) maxPost = 10 ;
-		pageList.setPageSize(maxPost) ;
+		pageList.setPageSize((int)maxPost) ;
 		UIForumPageIterator forumPageIterator = this.getChild(UIForumPageIterator.class) ;
 		forumPageIterator.updatePageList(pageList) ;
 		
@@ -138,7 +138,7 @@ public class UIViewTopic extends UIForm implements UIPopupComponent {
 		if(this.pageList == null) return null ;
 		UIForumPageIterator forumPageIterator = this.getChild(UIForumPageIterator.class) ;
 		this.pageSelect = forumPageIterator.getPageSelected() ;
-		long availablePage = this.pageList.getAvailablePage() ;
+		int availablePage = this.pageList.getAvailablePage() ;
 		if(this.pageSelect > availablePage) {
 			this.pageSelect = availablePage ;
 			forumPageIterator.setSelectPage(availablePage);
